step2/requirements.txt
google-adk
litellm
uv
google-adk[extensions]